+++ test/on_upgrade_test.dart
@@ -1,29 +1,10 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:on_upgrade/on_upgrade.dart';
-import 'package:pub_semver/pub_semver.dart';
 
 import 'utils/mock.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
-
-  test('Get last version', () async {
-    final onUpgrade = OnUpgrade();
-    mockSharedPreferences(<String, dynamic>{onUpgrade.keyLastVersion: '0.0.1'});
-
-    var result = await onUpgrade.getLastVersion();
-
-    expect(result, Version(0, 0, 1));
-  });
-
-  test('Get current Version', () async {
-    final onUpgrade = OnUpgrade();
-    mockPackageInfo(appName: 'test', packageName: 'package.test', version: '0.0.1', buildNumber: '1');
-
-    var result = await onUpgrade.getCurrentVersion();
-
-    expect(result, Version(0, 0, 1));
-  });
 
   test('Is initial', () async {
     final onUpgrade = OnUpgrade();
@@ -41,6 +22,16 @@ void main() {
     var result = await onUpgrade.isInitialInstallation();
 
     expect(result, false);
+  });
+
+  test('Is unknown', () async {
+    final onUpgrade = OnUpgrade();
+    mockSharedPreferences(<String, dynamic>{onUpgrade.keyLastVersion: '0.0.1'});
+
+    var result = await onUpgrade.isNewVersion();
+
+    expect(result.state, UpgradeState.unknown);
+    expect(result.hasError, true);
   });
 
   test('Is upgrade', () async {
@@ -71,6 +62,24 @@ void main() {
     expect(result, true);
   });
 
+  test('Get last version', () async {
+    final onUpgrade = OnUpgrade();
+    mockSharedPreferences(<String, dynamic>{onUpgrade.keyLastVersion: '0.0.1'});
+
+    var result = await onUpgrade.getLastVersion();
+
+    expect(result, '0.0.1');
+  });
+
+  test('Get current Version', () async {
+    final onUpgrade = OnUpgrade();
+    mockPackageInfo(appName: 'test', packageName: 'package.test', version: '0.0.1', buildNumber: '1');
+
+    var result = await onUpgrade.getCurrentVersion();
+
+    expect(result, '0.0.1');
+  });
+
   test('Custom lookup / update last version', () async {
     Future<bool> update([String version]) async => true;
     Future<String> lookup() async => '1.0.0';
@@ -79,7 +88,7 @@ void main() {
     var lookupResult = await onUpgrade.getLastVersion();
     var updateResult = await onUpgrade.updateLastVersion();
 
-    expect(lookupResult, Version(1, 0, 0));
+    expect(lookupResult, '1.0.0');
     expect(updateResult, true);
   });
 }