+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.0
+
+- **Breaking change:** `UpgradeWrapper.lastVersion`, `UpgradeWrapper.currentVersion`, `getLastVersion()` and `getCurrentVersion()` now return Strings.
+- Remove pub_semver dependency
+- Update documentation
+- Fix pub score issues
+- Clean up and extend tests
+
 ## 0.0.3
 
 - Readme fixes